requirements.txt
tensorflow==1.14.0  # CPU Version of TensorFlow.
# tensorflow-gpu==1.14.0  # GPU version of TensorFlow.
absl-py==0.8.1
astor==0.8.0
bert-tensorflow==1.0.1
gast==0.3.2
google-pasta==0.1.7
grpcio==1.24.3
h5py==2.10.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
numpy==1.17.3
pkg-resources==0.0.0
protobuf==3.10.0
scipy==1.3.1
six==1.12.0
tb-nightly==1.14.0a20190603
tensorboard==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019042513
Werkzeug==0.16.0
wrapt==1.11.2
